Cleanup, use i18n and remove unused dependency
