+++ package.json
@@ -45,7 +45,6 @@
     "react-is": "^19.2.0",
     "react-responsive": "^10.0.1",
     "react-router-dom": "^7.9.4",
-    "react-simple-wysiwyg": "^3.4.1",
     "react-slick": "^0.31.0",
     "recharts": "^3.4.1",
     "slick-carousel": "^1.8.1",
