+++ public/locales/en/translation.json
@@ -15,12 +15,14 @@
   "timeOfDay": "Time of day",
   "submit": "Submit",
   "edit": "Edit",
+  "preview": "Preview",
   "editName": "Edit {{name}}",
   "delete": "Delete",
   "deleteConfirmation": "Are you sure you want to delete \"{{name}}\"?",
   "add": "Add",
   "close": "Close",
   "difference": "Difference",
+  "useMarkdownHint": "You can use basic Markdown to format the text: *italic*, **bold**, - list",
   "days": "Days",
   "all": "All",
   "lastYear": "Last Year",
